Prevent Non-Malicious Packets from Targeting HoneyPots
Since the way the code's timing works assumes HoneyPots will only be placed between waves, the button to place them has been disabled during waves.
A check has been added whenever a packet's destination is set, and if the packet is not flagged as malicious it finds a new destination when its current one is a HoneyPot.