Fixed Bug with Firewall Panel
There was a bug with the firewall panel which would cause your settings to get corrupted when changing the currently selected firewall.
Now when a firewall is selected it prevents the toggles from updating settings until the UI has been completely updated.
Additionally, the pink and blue color options were swapped in the UI, they have now been swapped in the code as well. The order of operations as they appear in the code is the same as the order they appear in the UI.