Added BlackHatPlayer Manager
There is a new scene where the basic BlackHat UI is being tested, the BlackHat is capable of manipulating the properties of each starting point and the likelihood of a packet targeting one of the destinations.

All of the code common to both the BlackHat and WhiteHat bases has been factored out into a class called BassSharedBetweenHats.