Theming and Windows
Established a first theming pass. Changed all of the base prefabs to account for the new theming settings. Added windows (with no backend connection yet) for each of the intractable parts of the game.
Added a labeled slider prefab, which automatically updates it label when the slider is changed.
Added a script to assign toggle groups to toggles (we were having issues where the custom toggle scripts weren't getting recognized by their toggle groups.)