Added Player List
Instead of assuming that there will be exactly two players in a room, the Networking Manager now keeps track of a list of Networking.Players which track a photon player and some additional state about the player's role.
Supplemental utilities to support this transition have also been added.
The maximum number of players that can be in a room has been increased to 16.