Updated Gameplay to Support Multiple Players Per Side
Fixed the code when players join/leave (there wasn't a photon view on the networking manager in the gameplay scene).
Added suggested firewalls which whitehat advisors can manipulate to suggest firewall placement.
Advisors have alternative interactions (with a larger focus on simply seeing information as opposed to interacting with it).